+++ lib/src/message.dart
@@ -1,5 +1,6 @@
 import 'dart:typed_data';
 
+import 'message_prototype.dart';
 import 'messages_prototypes.dart';
 
 abstract class Message {
@@ -8,7 +9,7 @@ abstract class Message {
   /// encode `message` into bytes `Uint8List`
   ///
   /// if no `IMessagePrototype` defined for the `message` type an error will be thrown
-  static Uint8List encode(Message message) {
+  static Uint8List encode<T extends Message>(T message) {
     final bytes = MessagesPrototypes.instance.encodeMessage(message);
     if (bytes == null) {
       throw Exception(
@@ -16,6 +17,18 @@ abstract class Message {
       );
     }
     return bytes;
+  }
+
+  /// check if a [IMessagePrototype]  is defined for the `message` type [T]
+  static bool hasPrototype<T extends Message>() {
+    return MessagesPrototypes.instance.findPrototypeByMessageType<T>() != null;
+  }
+
+  /// define a new [prototype] for message of type [T]
+  ///
+  /// if [prototype] is already defined, it will be replaced
+  static void define<T extends Message>(IMessagePrototype<T> prototype) {
+    MessagesPrototypes.instance.definePrototype(prototype);
   }
 
   /// try to decode `bytes` into a `Message` object but only if the `bytes` are a valid message