+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.2.1
+
+- Fix MessagesPrototypes.encode: encode using incorrect IMessagePrototype
+
+- Add Message.hasPrototype, Message.definePrototype
+
+# 1.2.0
+
+- Fix BytesWriter: writes unsigned integers with reversed bytes
+
+- Add isEqual method at BytesReader
+
 # 1.1.0
 
 - Update license